backend/apps/staff/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User


class StaffSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ["id", "username", "email", "is_staff", "is_active"]
